Linter is drunk

+++ internal/grid/handlers.go
@@ -210,7 +210,7 @@ const (
 func init() {
 	// Static check if we exceed 255 handler ids.
 	// Extend the type to uint16 when hit.
-	if handlerLast > 255 {
+	if uint32(handlerLast) > 255 {
 		panic(fmt.Sprintf("out of handler IDs. %d > %d", handlerLast, 255))
 	}
 }
@@ -601,11 +601,13 @@ func GetCaller(ctx context.Context) *RemoteClient {
 
 // GetSubroute returns caller information from contexts provided to handlers.
 func GetSubroute(ctx context.Context) string {
+	//lint:ignore SA1029
 	val, _ := ctx.Value(ctxSubrouteKey{}).(string)
 	return val
 }
 
 func setCaller(ctx context.Context, cl *RemoteClient) context.Context {
+	//lint:ignore SA1029
 	return context.WithValue(ctx, ctxCallerKey{}, cl)
 }
 
@@ -681,6 +683,7 @@ func (h *StreamTypeHandler[Payload, Req, Resp]) NewRequest() Req {
 // These should be returned by the handler.
 func (h *StreamTypeHandler[Payload, Req, Resp]) PutRequest(r Req) {
 	if r != h.nilReq {
+		//lint:ignore SA6002
 		h.reqPool.Put(r)
 	}
 }
@@ -689,6 +692,7 @@ func (h *StreamTypeHandler[Payload, Req, Resp]) PutRequest(r Req) {
 // These should be returned by the caller.
 func (h *StreamTypeHandler[Payload, Req, Resp]) PutResponse(r Resp) {
 	if r != h.nilResp {
+		//lint:ignore SA6002
 		h.respPool.Put(r)
 	}
 }
